backend/requirements.txt
# ─── FastAPI & ASGI ──────────────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12

# ─── Validation & Paramètres ─────────────────────────────
pydantic==2.10.3
pydantic-settings==2.6.1

# ─── Base de données PostgreSQL ──────────────────────────
sqlalchemy==2.0.36
asyncpg==0.30.0

# ─── LangChain / LangGraph ───────────────────────────────
langchain==0.3.9
langchain-groq==0.2.1
langchain-community==0.3.9
langchain-core==0.3.21
langgraph==0.2.53
langsmith==0.1.147

# ─── RAG : ChromaDB + FastEmbed (ONNX, pas de torch) ────
chromadb==0.6.3
fastembed==0.8.0

# ─── Parsing CV ──────────────────────────────────────────
pymupdf==1.24.14
python-docx==1.1.2

# ─── Génération PDF rapport ──────────────────────────────
reportlab==4.2.5

# ─── Utilitaires ─────────────────────────────────────────
python-dotenv==1.0.1
loguru==0.7.3
tenacity==9.0.0
httpx==0.27.2
